% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{generate_circle}
\alias{generate_circle}
\title{Generates points on circle in terms of cartesian coordinates}
\usage{
generate_circle(r, x_center, y_center)
}
\arguments{
\item{r}{radius}

\item{x_center}{coordinate for x center}

\item{y_center}{coordinate for y center}
}
\value{
returns dataframe of x and y coordinates circle with desired parameters
}
\description{
Generates points on circle in terms of cartesian coordinates
}
\keyword{internal}
