% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbaanalysis-package.R
\docType{package}
\name{nbaanalysis-package}
\alias{nbaanalysis}
\alias{nbaanalysis-package}
\title{NBA Analysis}
\description{
A shiny app that allows users to select a currently active NBA player. The app provides an overview of the selected player on their career stats, heatmap of average shot accuracy by distance, as well as time-series plots to show career trends. Finally, the interface allows the user to choose two statistics to generate a scatterplot across all NBA players to view relationships between different stats. Users can filter by years of pro experience and position to narrow the scope and view how their selected player match up to the rest of the NBA.
}
\author{
\strong{Maintainer}: Johnny Antoun \email{johnny.antoun@duke.edu}

Authors:
\itemize{
  \item Eric Bai \email{eric.bai@duke.edu}
  \item Jackie Du \email{jacquelyn.du@duke.edu}
  \item Rich Fremgen \email{richard.fremgen@duke.edu}
}

}
\keyword{internal}
