% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{per_data}
\alias{per_data}
\title{Per game stats for 2016-2022 NBA season}
\format{
A data frame with 3580 rows and 28 variables
}
\usage{
data(per_data)
}
\description{
This table contains per-game data for NBA players from the 2016 to 2022 season
}
\details{
\itemize{
\item player. first and last name of player
\item age. player age
\item g. games played
\item gs games started
\item mp. minutes played per game
\item fg. field goal per game
\item fga. field goal attempted per game
\item pct_fg. percentage of field goals
\item 3p. 3 points field goals per game
\item 3pa. 3 points field goals attempts per game
\item pct_3p. 3 points field goals percentage
\item 2p.  2 points field goals per game
\item 2pa. 2 points field goals attempted per game
\item pct_2p. 2 point field goal percentage
\item pct_efg. Effective field goal percentage
\item ft. free throws per game
\item fta. free throws attempts per game
\item pct_ft. free throws percentage
\item orb. offensive rebounds per game
\item drb. defensive rebounds per game
\item trb. total rebounds per game
\item ast. assists per game
\item stl. steals per game
\item blk. blocks per game
\item tov. Turnovers per game
\item pf. personal fouls per game
\item pts. points per game
\item year. year of data
}
}
\keyword{datasets}
