% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{player_stats}
\alias{player_stats}
\title{Players' stats for 2016-2022 NBA season}
\format{
A data frame with 508 rows and 13 variables
}
\usage{
data(player_stats)
}
\description{
This table contains basic information about NBA players
}
\details{
\itemize{
\item name. first and last name of player
\item person_id. unique ID for each player in NBA
\item team_id. unique ID for team
\item jersey. jersey number
\item pos. player position
\item height_ft. number of feet in height
\item height_in. number of additional inches in height
\item weight_lbs. weight in pounds
\item dob. date of birth
\item years_exp. years of professional basketball experience
\item country. country of origin
\item team_name. full name of team
\item team_abbr. abbreviated name of team
}
}
\keyword{datasets}
