% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_functions.R
\name{plot_shooting_heatmap}
\alias{plot_shooting_heatmap}
\title{Use diagram of half court to create heatmap based on shooting accuracy from different distances to basket}
\usage{
plot_shooting_heatmap(
  dist_0_3,
  dist_3_10,
  dist_10_16,
  dist_16_3p,
  dist_3p,
  primary,
  secondary
)
}
\arguments{
\item{dist_0_3}{Shooting accuracy percentage between 0 and 3 feet}

\item{dist_3_10}{Shooting accuracy percentage between 3 and 10 feet}

\item{dist_10_16}{Shooting accuracy percentage between 10 and 16 feet}

\item{dist_16_3p}{Shooting accuracy percentage between 16 feet and 3-pt line}

\item{dist_3p}{Shooting accuracy percentage for 3-pts}
}
\value{
A ggplot heatmap of player scoring accuracy
}
\description{
Use diagram of half court to create heatmap based on shooting accuracy from different distances to basket
}
\keyword{internal}
