% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shooting_data}
\alias{shooting_data}
\title{Shooting stats for 2016-2022 NBA season}
\format{
A data frame with 3578 rows and 17 variables
}
\usage{
data(shooting_data)
}
\description{
This table contains shooting data for NBA players from the 2016 to 2022 season
}
\details{
\itemize{
\item player. first and last name of player
\item age. player age
\item g. games played
\item mp. minutes played
\item fg. field goal percentage
\item dist. average distance of field goal attempts
\item pct_of_fg_2p. percentage of field goals that are 2 pointers
\item pct_of_fg_0_3. percentage of field goals from 0-3 feet
\item pct_of_fg_3_10. percentage of field goals from 3-10 feet
\item pct_of_fg_10_16. percentage of field goals from 10-16 feet
\item pct_of_fg_16_3p.  percentage of field goals from 16 feet to 13
\item pct_of_fg_3p. percentage of field goals that are 3 pointers
\item fg_pct_2p. 2 point field goal percentage
\item fg_pct_0_3. percentage of field goals from 0-3 feet
\item fg_pct_3_10. percentage of field goals from 3-10 feet
\item fg_pct_10_16. percentage of field goals from 10-16 feet
\item fg_pct_16_3p. percentage of field goals from 16 feet   to 3 pointers
\item year. year of data
}
}
\keyword{datasets}
